requirements.txt
numpy==2.0.0
python-dateutil==2.9.0.post0
setuptools==69.5.1
six==1.16.0
wheel==0.43.0
